Text Formatting
This is a practice for text and button formatting.

This project doesn't aim to really achieve anything but
just to get me familiar with a few CSS syntax. Before the
project, some texts were not reflecting the desired colors
and size, but it is fixed now. Button background was also fixed.